+++ lib/scalingo/connection.rb
@@ -8,14 +8,11 @@ module Scalingo
     def build_connection(opts = {})
       raise MissingToken if !token
 
-      # Some requests override global client configuration, like JWT token exchanging
-      should_parse_json = opts[:always_json] || parse_json
-
       Faraday::Connection.new(connection_options) do |connection|
         connection.use Faraday::Request::Multipart
         connection.use Faraday::Request::UrlEncoded
-        connection.use Faraday::Response::ParseJson if should_parse_json
         connection.use FaradayMiddleware::RaiseHttpException
+        connection.response :json, :content_type => /\bjson$/
         connection.adapter(adapter)
       end
     end