+++ lib/scalingo/endpoint/addons.rb
@@ -2,12 +2,12 @@ module Scalingo
   module Endpoint
     class Addons < Collection
       def create(addon_provider_id, plan_id)
-        post(nil, {addon:{addon_provider_id: addon_provider_id, plan_id: plan_id}})
+        post(nil, addon: { addon_provider_id: addon_provider_id, plan_id: plan_id })
       end
     end
     class Addon < Resource
       def update(plan_id)
-        patch(nil, {addon: {plan_id: plan_id}})
+        patch(nil, addon: { plan_id: plan_id })
       end
 
       def destroy
@@ -16,4 +16,3 @@ module Scalingo
     end
   end
 end
-