+++ lib/scalingo/endpoint/collaborators.rb
@@ -2,7 +2,7 @@ module Scalingo
   module Endpoint
     class Collaborators < Collection
       def create(email)
-        post(nil, {collaborator: {email: email}})
+        post(nil, collaborator: { email: email })
       end
     end
     class Collaborator < Resource
@@ -12,4 +12,3 @@ module Scalingo
     end
   end
 end
-