+++ lib/scalingo/endpoint/deployments.rb
@@ -9,4 +9,3 @@ module Scalingo
     end
   end
 end
-