+++ lib/scalingo/endpoint/domains.rb
@@ -3,15 +3,15 @@ module Scalingo
     class Domains < Collection
       def create(name, tlscert = nil, tlskey = nil)
         if tlskey
-          post(nil, {domain: {name: name, tlscert: tlscert, tlskey: tlskey}})
+          post(nil, domain: { name: name, tlscert: tlscert, tlskey: tlskey })
         else
-          post(nil, {domain: {name: name}})
+          post(nil, domain: { name: name })
         end
       end
     end
     class Domain < Resource
       def update(tlscert, tlskey)
-        path(nil, {domain: {tlscert: tlscert, tlskey: tlskey}})
+        path(nil, domain: { tlscert: tlscert, tlskey: tlskey })
       end
 
       def destroy
@@ -20,4 +20,3 @@ module Scalingo
     end
   end
 end
-