+++ lib/scalingo/regions_cache.rb
@@ -6,7 +6,7 @@ module Scalingo
       region = @regions_cache.select { |r| r.name == name }.first
       raise InvalidRegion, name if region.nil?
 
-      return "https://#{region['api']}/v1"
+      return "#{region['api']}/v1"
     end
   end
 end